kharazmichi-bot/requirements.txt
# ===========================================
# KHARAZMICHI BOT - Dependencies
# ===========================================

# Web Framework
fastapi==0.115.0
uvicorn[standard]==0.30.0

# Telegram Bot
python-telegram-bot==21.5

# OpenAI & AI
openai==1.54.0

# Supabase
supabase==2.9.0

# Data Validation
pydantic==2.9.0
pydantic-settings==2.5.0

# Document Processing
python-docx==1.1.2
docx2txt==0.8

# Utilities
python-dotenv==1.0.1
httpx==0.27.0
aiofiles==24.1.0

# Logging
loguru==0.7.2
